build on systems with sizeof(int) < sizeof(void*)

+++ util.cpp
@@ -8,7 +8,7 @@ QString Util::displayString(const QObject* object)
 {
   if ( object->objectName().isEmpty() ) {
     return QString::fromLatin1( "0x%1 (%2)" )
-      .arg( QString::number( reinterpret_cast<int>( object ), 16 ) )
+      .arg( QString::number( reinterpret_cast<qulonglong>( object ), 16 ) )
       .arg( object->metaObject()->className() );
   }
   return object->objectName();
